events render. cleaning up events moved from Events to SingleEvent since SingleEvent is using the eventList.
