+++ src/components/app/events/event-actions.js
@@ -25,37 +25,6 @@ function loadingEventsFailed(error) {
   };
 }
 
-
-
-export const CLEARING_EVENTS = 'CLEARING_EVENTS';
-function clearingEvents(emptyList) {
-  return {
-    type: CLEARING_EVENTS,
-    loading: true,
-    eventList: emptyList,
-  };
-}
-
-// export const CLEARING_EVENTS_SUCCEEDED = 'CLEARING_EVENTS_SUCCEEDED';
-// function clearingEventsSucceeded() {
-//   return {
-//     type: CLEARING_EVENTS_SUCCEEDED,
-//     loading: true,
-//     isError: false,
-//   };
-// }
-
-// export const CLEARING_EVENTS_FAILED = 'CLEARING_EVENTS_FAILED';
-// function clearingEventsFailed(error) {
-//   return {
-//     type: CLEARING_EVENTS_FAILED,
-//     isError: false,
-//     error: error,
-//   };
-// }
-
-
-
 export function loadUserEvents(id) {
   return function(dispatch) {
     dispatch(loadingEvents());
@@ -95,15 +64,5 @@ export function loadLocationEvents(id) {
     if (isError) {
       dispatch(loadingEventsFailed('error message'));
     }
-  };
-}
-
-export function clearEvents() {
-  return function(dispatch) {
-
-    dispatch(clearingEvents([]));
-
-    // check
-
   };
 }
